services/style-analysis/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.3
hypothesis==6.92.2
